requirements.txt
nltk==3.5
pandas==1.3.0
requests==2.22.0
plotly==5.1.0
scipy==1.4.1
dash_bootstrap_components==0.12.2
twint==2.1.20
matplotlib==3.1.2
dash_html_components==1.1.4
seaborn==0.11.1
dash==1.21.0
dash_core_components==1.17.1
numpy==1.18.1
python_dateutil==2.8.2
scikit_learn==0.24.2
